+++ lib/core.dart
@@ -4,7 +4,9 @@ import 'package:flutter/material.dart';
 
 import 'src/utils/data.dart' as data;
 
-export 'src/errors/error_pages.dart';
+export 'src/errors/http_error_pages.dart';
+export 'src/errors/models.dart';
+export 'src/errors/pages/http_error_page.dart';
 export 'src/utils/data.dart';
 export 'src/utils/extensions.dart';