+++ lib/src/errors/models.dart
@@ -1,0 +1,2 @@
+export 'models/app_exception.dart';
+export 'models/http_response_exception.dart';