+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## 1.2.1
+
+### Changed
+
+- `ErrorPage` is deprecated. Use `HttpErrorPage` instead.
+- `ErrorPages` is deprecated. Use `HttpErrorPages` instead.
+- The `name` property is removed in the `HttpErrorPage`. But the error name will
+  be customized based on the `statusCode` property and displayed on the page.
+- `HttpErrorPages.byCode()` is deprecated. Use `HttpErrorPage` directly instead.
+
+### Added
+
+- `AppException` (Default implementation of `Exception`. But the message is not
+  private).
+- `CoreLocalizations` (Localization for this package).
+- `HttpResponseException` (Data model for `HttpErrorPage`).
+
+  So `HttpErrorPage` can now be created from `HttpResponseException`.
+
+  ```dart
+  HttpErrorPage.fromException(
+    HttpResponseException(...),
+  );
+  ```
+
+- `HttpErrorPage` will have a default message that varies according to the given
+  status code. The messages are also localized if you include
+  `CoreLocalizations.delegate` to your app.
+
+### Fixed
+
+- Changing the `..addAll()` method on `Map` to spread operator method.
+
 ## 1.1.0
 
 ### Added